Refactor code and fix bugs
