--- src/backend/langflow/base/models/__init__.py
@@ -1,3 +1,0 @@
-from .model import LCModelComponent
-
-__all__ = ["LCModelComponent"]
