+++ src/backend/langflow/components/tools/SearchAPITool.py
@@ -1,7 +1,6 @@
 from langchain_community.tools.searchapi import SearchAPIRun
 from langchain_community.utilities.searchapi import SearchApiAPIWrapper
-
-from langflow import CustomComponent
+from langflow.custom import CustomComponent
 from langflow.field_typing import Tool
